Playing with methods - Constrictor accepts reguest handler,     - cause class HttpWebServer don't cares what and how will be processed     - his task start server for listening port - Implements readonly fields     - port     - server
